+++ frontend/src/components/Navigation/ProfileButton.js
@@ -1,13 +1,10 @@
 import React, { useState, useEffect } from "react";
 import { useDispatch } from "react-redux";
-import { useHistory } from "react-router-dom";
 import * as sessionActions from "../../store/session";
 
 function ProfileButton({ user }) {
   const dispatch = useDispatch();
   const [showMenu, setShowMenu] = useState(false);
-
-  const history = useHistory();
 
   const openMenu = () => {
     if (showMenu) return;
@@ -31,11 +28,6 @@ function ProfileButton({ user }) {
     dispatch(sessionActions.logout());
   };
 
-  const addRestaurant = e => {
-    e.preventDefault();
-    history.push("/businesses/new");
-  };
-
   return (
     <>
       <button onClick={openMenu}>
@@ -45,9 +37,6 @@ function ProfileButton({ user }) {
         <ul className="profile-dropdown">
           <li>{user.username}</li>
           <li>{user.email}</li>
-          <li>
-            <button onClick={addRestaurant}>Create a restaurant</button>
-          </li>
           <li>
             <button onClick={logout}>Log Out</button>
           </li>
